+++ CHANGELOG.md
@@ -51,6 +51,7 @@
 * Add option to remove leading zeroes from Inventory HRIDs. Refs UIIN-1398
 * Add support for maintaining copy-cataloguing profiles in the settings area. Towards UIIN-1401. Entails adding new interface `copycat-profiles`.
 * Add support for importing MARC records from external Z39.50 sources. Towards UUIN-1317 and UUIN-1318. Entails adding new interface `copycat-imports`.
+* Parse item.enumeration numerically for sorting if possible. Refs UIIN-1412.
 
 ## [5.0.1](https://github.com/folio-org/ui-inventory/tree/v5.0.1) (2020-10-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v5.0.0...v5.0.1)